rpk: move cmd/generate to its own package
