+++ src/go/rpk/pkg/cli/cmd/generate/generate.go
@@ -7,21 +7,20 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-package cmd
+package generate
 
 import (
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cmd/generate"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
 
-func NewGenerateCommand(fs afero.Fs) *cobra.Command {
+func NewCommand(fs afero.Fs) *cobra.Command {
 	command := &cobra.Command{
 		Use:   "generate [template]",
 		Short: "Generate a configuration template for related services",
 	}
-	command.AddCommand(generate.NewGrafanaDashboardCmd())
-	command.AddCommand(generate.NewPrometheusConfigCmd(fs))
-	command.AddCommand(generate.NewShellCompletionCommand())
+	command.AddCommand(newGrafanaDashboardCmd())
+	command.AddCommand(newPrometheusConfigCmd(fs))
+	command.AddCommand(newShellCompletionCommand())
 	return command
 }
